feat: three number sum
It takes in a non-empty array of distinct integers and an integer
representing a target sum. The function finds all triplets in the
array that sum up to the target sum and return a two-dimensional array of all these triplets. The numbers in each triplet are ordered in ascending order, and the triplets themselves are ordered in ascending order with respect to the numbers they hold.
They function returns an empty array if no three numbers sum up to the target sum.