feat: smallest difference algorithm
Take in two non-empty arrays of integers, find the pair of numbers (one from each array) whose absolute difference
is closest to zero, and return an array containing those two
numbers, with the number from the first array in the first
position.
Assume there will be only one pair of numbers with the smallest difference.