feat: spiral traverse algorithm
Takes in a n x m two-dimensional array (that can be square-shaped when n == m) and returns a one-dimensional array of all the array's elements in spiral order.

Spiral order starts at the top left corner of the two-dimensional array, goes to the right, and proceeds in spiral pattern all the way until every element has been visited.