feat: structure of min number of coins for change algo
Given an array of positive integers representing coin denominations  and a single non-negative integer n representing a target amount of money, write a function that returns the smallest number of coins needed to make change for (to sum up to) that target amount  using the given coin denominations.

You have an unlimited amount of coins.If it's impossible to make change for the target amount, return -1.